+++ src/parquet/schema/types.h
@@ -253,11 +253,11 @@ class GroupNode : public Node {
 
   virtual bool Equals(const Node* other) const;
 
-  const NodePtr& field(size_t i) const {
+  const NodePtr& field(int i) const {
     return fields_[i];
   }
 
-  size_t field_count() const {
+  int field_count() const {
     return fields_.size();
   }
